requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.43.0
django-cors-headers==3.5.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.10.3
djangorestframework-simplejwt==4.5.0
djoser==2.1.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
dj-rest-auth==2.1.1
uritemplate==3.0.1
urllib3==1.25.11
zipp==3.4.0
